refactor: addComma 함수를 파일로 따로 분리
