+++ OpenMarket/OpenMarket/ViewController/GridViewController.swift
@@ -14,7 +14,7 @@ class GridViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        openMarketAPIManager.requestProductList(of: 1) { (result) in
+        openMarketAPIManager.requestProductList(of: 7) { (result) in
             switch result {
             case .success (let product):
                 self.productList.append(contentsOf: product.items)
